[MODIFY] query 파일 생성 및 구조 변경
