+++ api/controllers/itemsController.js
@@ -1,12 +1,6 @@
 const itemsService = require("../services/itemsService");
 const { asyncErrorHandler } = require("../middleware/errorHandling");
 
-const sortCategory = {
-  new: "created_at DESC",
-  expensive: "price DESC",
-  cheap: "price",
-};
-
 const getItem = asyncErrorHandler(async (req, res) => {
   const { itemId } = req.params;
   const item = await itemsService.getItem(itemId);
@@ -15,22 +9,19 @@ const getItem = asyncErrorHandler(async (req, res) => {
 });
 
 const getItemsList = asyncErrorHandler(async (request, response) => {
-  //DEFAULT ORDER WOULD BE created_at DESC
-  let orderString = sortCategory[request.query.sort]
-    ? sortCategory[request.query.sort]
-    : "created_at DESC";
+  //GETTING SORT CONDITION
+  const sort = !request.query.sort ? "default" : request.query.sort;
 
   //GETTING CATEGORY PAGE BY NUMBER
-  let category = request.query.category;
+  const category = request.query.category;
 
-  //DEFAULT PAGE NUMBER WOULD BE 1
-  let pageNumber = !request.query.page ? 1 : request.query.page;
+  //DEFAULT PAGE NUMBER IS 1
+  const page = !request.query.page ? 1 : request.query.page;
 
-  const result = await itemsService.getItemsList(
-    orderString,
-    category,
-    pageNumber
-  );
+  //GETTING SEARCH STRING
+  const search = request.query.search;
+
+  const result = await itemsService.getItemsList(sort, category, page, search);
   return response.status(200).json({ data: result });
 });
